Copy Intel key into Docker image during `gsc build-gramine`
Copy Intel repo signing key into the resulting base-Gramine Docker image
(as part of the `gsc build-gramine` command). This step was missing and
resulted in failure of `gsc build-gramine`. Additionally, this commit
improves the code quality of the `build-gramine` logic and adds a check
for `gsc build` on incorrectly specified base-Gramine Docker image in
the config file.

Co-authored-by: manju956 <manjunath.a.c@intel.com>
Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>
Signed-off-by: manju956 <manjunath.a.c@intel.com>