Work around TOML Python parser bug of parsing `\x`
The currently used TOML parser in Python is buggy and doesn't allow to
parse filenames like `system-systemd\x2dcryptsetup.slice`. Since the
only files that we detected in real world are systemd helper files
(which are not used by graminized apps anyway), the current workaround
is to simply skip such files when collecting `sgx.trusted_files`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>